imapmda: don't file the same message multiple times in the same folder
With multiple rules without stop or keep directives, it's possible
for a message to be assigned to the same folder multiple times.  This
can result in duplicate messages in the target folder.

This commit deduplicates the delivery rules based on the folder
and flags of the message.  If multiple flag sets are used, the message
will still be delivered multiple times with different flags.  This is
intentional as the flag set can be modified over the course of the
rule set.